requirements.txt
kagglehub==1.0.0
requests>=2.32
tqdm>=4.67
PyYAML>=6.0
scikit-learn>=1.3
xgboost>=2.0
pandas>=2.0
numpy>=1.24
joblib>=1.3
matplotlib>=3.7
seaborn>=0.12
google-genai>=1.0
gradio>=4.0
pillow>=10.0
python-dotenv>=1.0
